+++ src/d3-funnel/d3-funnel.js
@@ -251,7 +251,7 @@ class D3Funnel {
 				fill: this.colorizer.getBlockFill(block, index, this.fillType),
 				label: {
 					raw: label,
-					formatted: this.labelFormatter.format(label, count),
+					formatted: this.labelFormatter.format(label, block[1]),
 					color: this.colorizer.getLabelFill(block),
 				},
 			});